Move asset decorators from assets to organization.decorators
The four asset access-control decorators (asset_is_recorder,
asset_is_operator, asset_is_owner, asset_id_in_get_post) belong in
organization since their logic depends on org-role helpers. This breaks
the assets → organization import in assets/decorators.py.

assets/decorators.py is kept as a re-export shim for now.
